CI Minutes first draft + ultimate parent id as int in upstream tables

+++ transform/snowflake-dbt/models/legacy/gitlab_dotcom/staging/gitlab_dotcom_namespace_lineage_prep.sql
@@ -41,7 +41,7 @@ WITH RECURSIVE namespaces AS (
 
   SELECT
     *,
-    GET(upstream_lineage, ARRAY_SIZE(upstream_lineage)-1) AS ultimate_parent_id -- Last item is the ultimate parent.
+    GET(upstream_lineage, ARRAY_SIZE(upstream_lineage)-1)::INT   AS ultimate_parent_id -- Last item is the ultimate parent.
   FROM recursive_namespaces
 
   UNION ALL
